Adds jedi-mode to list of unsupported minor modes
jedi-mode is a minor mode implemented by jedi.el, a Python auto-complete
package built on top of Jedi and auto-complete.el. The two primary
features of jedi-mode are:
1) Automatically start autocompletion when the dot key is pressed
2) Activate key bindings for moving to variable and function
definitions.

Neither make much sense in the context of multiple cursors. In fact,
because multiple-cursors disables auto-complete-mode, jedi-mode's
complete-on-dot function will fail to run correctly. The end result is
that typing a dot with multiple cursors active results in only one
cursor producing the dot.

Because of this bug, and that jedi-mode as a whole doesn't make much
sense with multiple cursors, I have added jedi-mode to the list of
unsupported minor modes.