Evil tests throw error if cmd not in cmds to run [once|all]
I don't think a user's personal commands list file should be loaded when running
tests, a few tests broke for a command that was in my personal `run once list`
instead of being in the `run for all` list.

It was grievous to debug.

This way it doesn't happen again :+1:

Additionally this is actually pretty cool for evil commands as adding new
functionality or new tests or debugging etc.

- `mc-evil.el`

  * Added a predicate that says we are using `evil-mode`

- `env.el`

  * Advice around `mc/prompt-for-inclusion-in-white-list` that is a pass through if
    not using evil in the tests, and if using evil, throws an error that the command
    isn't in one of the two default command variables.

Previously, if some command wasn't in those lists, the whole test would just
hang waiting for user input that never came.

Now, it outputs the message, and returns nil because of the wrapped
`ignore-errors` in `mc/execute-this-command-for-all-cursors`.

When this happens, you can see the commands that didn't make it, and add them to
the appropriate list and re-run the test :smile_cat: