:white_check_mark: evil state change tests working :ghost:
We are keeping track of state changes when restoring overlay variables for
cursors.  This is *only for fake cursors* and only when using evil.

Basically, for the fake cursor, go to its point and mark, then restore the
rest of the state, then change the state if you need to.

There is one odd behavior going from `insert` state to `normal` state.  This
happens when the main cursor goes from `insert` to `normal`, then the fake
cursor is in `normal` state and the command to execute will take it to normal.

Since this executes after the main cursor, we need to change to `normal`
state, but the way evil handles point and mark, we *dont* want the cursor
moved back like it does, so we disable it and set it back to its original
value to avoid that and keep all the point and marks proper.

+ see `mc/evil-restore-state-from-overlay` `mc/restore-state-from-overlay`

+ `mc/restore-point-and-mark-from-overlay` is refactored out to be called in
different places now.

+ Keep track of new evil vars to save per cursor
  - see `mc-vars.el` and `mc--evil-cursor-specific-vars`
  - `evil-state` is obvious, but `evil-previous-state`,
    `evil-previous-state-alist`, and `evil-next` state are global variables
    set in the evil macros and need to be kept track of.
  - `evil-normal-state-entry-hook` needs to be kept track of because
    `evil-visual-state` sets this hook when exiting `visual` state into
    `insert` state.

+ New commands to run for all cursors
  - see 'mc-vars.el` and `mc--evil-cmds-to-run-for-all`
  - every state toggle is added and commands that take you into a new state
  - `evil-delete-char` is used in the tests and needs to be added to pass them

+ evil specific vars and commands are only appended to the default vars and
commands when using evil *only*