:red_circle: evil-repeat fixes but getting `End of line` error :boom:
`evil-repeat-pre-hook` gets advice wrapped around it to only get called when not
executing for a fake cursor.  Much like the other advises written.

We only need to record the command for repetition just the once for the main
cursor.  The other commands will use that same repetition so it is all good:+1:.

`evil-repeat-ring` needs to be stored on a per cursor basis to get the right
repeat command.

`evil-repeat` command needs to be added to the list of commands to run for all
cursors.