:white_check_mark: mark-next-like-this supported using evil :ghost:
mark functions will leave *all* cursors in `visual` state to perform further
operations on.

`mc/keyboard-quit` transitions cursors into `normal` state or if in visual state
use `evil-exit-visual-state` which goes to the `evil-previous-state`.

When marking things and creating cursors from them, a user using evil would have
to mark something, going in to `visual` state.  The mark and point attained by
searching some regex forward and pushing the mark and point coincide with evil's
"expanded-region".  That is the state where emacs commands work like normal.
`evil-visual-mark` and `evil-visual-point` are contracted point and mark.

The main cursor is in `visual` state, `evil-visual-region-expanded` is true for
the main cursor **and** the fake cursors will have set a point and mark that
would be "expanded" like the main cursor.

When the mark command finishes, the `evil-visual-post-command` will run and
"contract" the region.  So we can use this information to emulate the fake
cursor's point and mark by calling 'evil-visual-contract-region` before storing
the state to the overlay.

Because the main cursor is still "expanded" and because
`evil-visual-contract-region` sets `evil-visual-region-expanded` to `nil` we
need to reset it back to `t` in case there are more cursors.

Restoring state for a fake cursor during a transition needs a special flag
`evil-visual-contracted` so that when we restore state we can expand the point
and mark after changing the state so that commands will work on the proper
region.

After storing in visual state and contracting region, need to re-set point and
mark from `evil-visual-beginning` and `evil-visual-end` based off of point and
mark direction.  This ensures the searching forward for the next regex is after
the last search so nothing cycles.