:red_circle: testing mark-all-dwim :boom:
One test is failing.  Marking a single letter goes into
'mark-all-symbols-like-this-in-defun` which finds a single letter and then keeps
setting point one value previous and it keeps going in an infinite loop.

Strange because it just goes out to `mark-all-like-this` and that test finds the
single letter just fine.

I think it is the symbol boundary being placed in the `regexp-opt` but i'm not
sure.