:warning: :100: evil-delete test is passing with some caveats
- "dd" doesn't work, only "d$" and "D"

mark-all-dwim fixed to handle various times going in and out of visual state.

need to exit visual state if num cursors is one and we are exiting mc-mode, the
mark or the char needs to be adjusted after exiting visual state just in case we
keep going for a search.