:white_check_mark: evil-insert tests passing :ghost:
`evil-maybe-remove-spaces` is set to `t` when transitioning to insert state
programatically.  This is odd because my setup has this variable set to `nil` in
every state.

With some proper debug output, can see the value is originally `nil` and on the
transition to `insert` state it is set to `t`.

This was causing the command `evil-normal-state`, which calls
`evil-maybe-remove-spaces`, to not hit the first condition to remove the command
from the post command hook only -- instead going with `do-remove` condition.