fix for using evil-text-objects
These motions operate with the evil interactive code `<v>`

That means it uses the current visual selection and their variables to perform
the motion.

more often than not, those variables don't match where point and mark are, so
for these types of commands and these *only* we need to refresh with
`evil-visual-refresh`.  This includes the main cursor.

This could be brought up in a pre command hook too I suppose, but for now it is
in `make-a-note-of-command-being-run` and `save-excursion`.

`save-excursion` needs it before the main cursor state is restored, otherwise
the same thing we are trying to prevent happens again.