:white_check_mark: evil-change tests fix passing :ghost:
fix for evil visual line from evil-change

selecting a visual line produces the right ranges, but changing the visual
selection sends us into insert state for the main cursor, and we change states
back to visual state, before we change the selection.

This is problematic because changing to visual state sets the
evil-visual-selection to `char` not to `line` like it was.  need to refresh with
the same mark and point and make sure the type is passed in as `line` and
everything works.