:white_check_mark: mark-all-like-this-dwim tests passing :ghost:
searching using dwim for a single char works now

when storing a visual region, need to cache original point and mark for the
direction evaluation, moving point and mark.  When contracting a visual region,
evil always sets point and mark to be `point > mark` no matter what it was
before.