Mention NonGNU ELPA
Hi,

multiple-cursors has recently been added to [NonGNU ELPA](https://elpa.nongnu.org/), allowing Emacs 28+ users to install it without any additional configuration.

This patch mentions NonGNU in the README, so that users know that it can be installed by default.

Otherwise, the only thing worth noting is that NonGNU ELPA distributes stable versions by default, and these are identified by commits that modify the version tag in the package header. I'm not sure if new versions are to be expected, but in case there is an update, please remember to update the version tag too.